requirements.txt
tensorflow
flask
streamlit
pillow
requests
numpy